requirements.txt
discord.py>=1.3.4,<2
tomlkit>=0.6,<1
mysqlclient>=2.0.1,<3
